feat: 3대력관리자부여 post기능추가
